Remove all calls to unwrap in client code
This makes it possible for callers to handle errors themselves.
